+++ install.sh
@@ -2,5 +2,5 @@
 # Get script file dir
 script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-pip uninstall -y auraview
-pip install "$script_dir"
+uv tool uninstall auraview
+uv tool install "$script_dir"